requirements.txt
requests
Flask==3.0.3
gunicorn==20.1.0
langchain==0.2.11
langchain-groq==0.1.6
python-dotenv
gdown==4.6.0
feedparser
faiss-cpu==1.7.4
pypdf2==3.0.1
feedparser==6.0.10
sentence-transformers==2.2.2
numpy==1.24.3
tqdm==4.65.0